requirements.txt
flask==2.2.3
python-dotenv==1.0.0
flask-restful==0.3.9
flask_sqlalchemy==3.0.3
mysqlclient==2.1.1
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
Werkzeug~=2.2.3
marshmallow~=3.16.0